add enpoint to json, remove duplicate function, organise functions into better fitting files
