+++ controllers/comments.controller.js
@@ -1,4 +1,51 @@
-const { removeCommentById, updateCommentVotes } = require("../models/comments.model");
+const {
+  selectArticleComments,
+  createComment,
+  removeCommentById,
+  updateCommentVotes,
+} = require("../models/comments.model");
+const { selectArticlesById } = require("../models/articles.model");
+const { selectSingleUser } = require("../models/users.model");
+
+const getArticleComments = (req, res, next) => {
+  const articleId = req.params.article_id;
+  const promises = [selectArticleComments(articleId)];
+  {
+    promises.push(selectArticlesById(articleId));
+  }
+
+  Promise.all(promises)
+    .then((resolvedPromises) => {
+      res.status(200).send({ comments: resolvedPromises[0] });
+    })
+    .catch((err) => {
+      next(err);
+    });
+};
+
+const postArticleComment = (req, res, next) => {
+  const articleId = req.params.article_id;
+  const username = req.body.username;
+  const commentObj = req.body;
+  const commentBody = req.body.body;
+  if (commentBody) {
+    selectSingleUser(username)
+      .then(() => {
+        return selectArticlesById(articleId);
+      })
+      .then(() => {
+        return createComment(articleId, commentObj);
+      })
+      .then((comment) => {
+        return res.status(201).send({ comment });
+      })
+      .catch((err) => {
+        next(err);
+      });
+  } else {
+    res.status(400).send({ msg: "No comment given" });
+  }
+};
 
 const deleteCommentById = (req, res, next) => {
   const commentId = req.params.comment_id;
@@ -23,4 +70,9 @@ const patchCommentVotes = (req, res, next) => {
     });
 };
 
-module.exports = { deleteCommentById, patchCommentVotes };
+module.exports = {
+  getArticleComments,
+  postArticleComment,
+  deleteCommentById,
+  patchCommentVotes,
+};
