+++ controllers/topics.controller.js
@@ -6,5 +6,4 @@ const getTopics = (req, res, next) => {
   });
 };
 
-
 module.exports = { getTopics };
