+++ endpoints.json
@@ -25,6 +25,26 @@
       ]
     }
   },
+  "POST /api/articles": {
+    "description": "posts a new article as long as user and topic are on the system",
+    "queries": [],
+    "exampleResponse": {
+      "articles": [
+        {
+          "author": "butter_bridge",
+          "title": "My top ten summer recipes",
+          "body": "1. Mango chicken skewers, 2. Halloumi burgers, 3. Ceaser salad, 4. The best bruscetta, 5. Persian orange and cinamon rice salad, 6. Curried tofu skewers, 7. Prawn and pineapple tacos, 8. Eton mess, 9. Quick pad thai, 10. Spicy mexican bean burgers",
+          "topic": "mitch",
+          "article_img_url": "example-image-url",
+          "article_id": 3,
+          "votes": 0,
+          "created_at": "2018-05-30T15:59:13.341Z",
+          "comment_count": "0"
+        }
+      ]
+    }
+  },
+
   "GET /api/articles/:article_id": {
     "description": "serves an object of specific requested article",
     "queries": ["sort_by", "order"],
