+++ models/comments.model.js
@@ -1,5 +1,33 @@
 const db = require("../db/connection");
 
+const selectArticleComments = (articleId) => {
+  return db
+    .query(
+      `SELECT comment_id,votes, created_at, author, body, article_id 
+      FROM comments
+      WHERE article_id = $1
+      ORDER BY created_at DESC`,
+      [articleId]
+    )
+    .then((result) => {
+      return result.rows;
+    });
+};
+
+const createComment = (articleId, commentObj) => {
+  const { username, body } = commentObj;
+  return db
+    .query(
+      `INSERT INTO comments (body, article_id, author)
+      VALUES ($1, $2, $3)
+      RETURNING body`,
+      [body, articleId, username]
+    )
+    .then((result) => {
+      return result.rows[0].body;
+    });
+};
+
 const removeCommentById = (commentId) => {
   return db
     .query(`DELETE FROM comments WHERE comment_id =$1 RETURNING body`, [
@@ -33,4 +61,9 @@ const updateCommentVotes = (newCommentVotes, commentId) => {
     });
 };
 
-module.exports = { removeCommentById, updateCommentVotes };
+module.exports = {
+  selectArticleComments,
+  createComment,
+  removeCommentById,
+  updateCommentVotes,
+};
