+++ models/topics.model.js
@@ -6,4 +6,22 @@ const selectTopics = () => {
   });
 };
 
-module.exports = { selectTopics };
+const checkTopicExists = (topic) => {
+  return db
+    .query(
+      `SELECT * FROM topics
+    WHERE slug = $1`,
+      [topic]
+    )
+    .then(({ rows }) => {
+      if (!rows.length) {
+        return Promise.reject({
+          status: 404,
+          msg: "Topic not found",
+        });
+      }
+      return rows[0];
+    });
+};
+
+module.exports = { selectTopics, checkTopicExists };
